. Changed the variable names to match convention (example: JButtons will include button in the name, etc.)
Implemented a window listener to show a farewell screen when the user
exits the program.

Added a file reader system to scan in the levels:
* Allows easy editing of the levels.
* Allows me to access the levels whenever I want.
* Keeps code form getting too messy.

Added keyListener that listens to the arrow keys --> not added to any
component yet.

Made an array to track the motion of the ball.

Added timers to change the add motion --> changes the ball array.

Created a paint method -- not actually working yet.