. Added repainting statement in the timer.
Added try catch statements to loop around the board array so that the
ball can move from one side to another.

Changed the level format to not include new lines. This makes it easier
to read into the program.

Added KeyListener to the gameFrame and set Focusables.

Changed arrays to static arrays for easier access from the other class.