. Added collision detection so the ball no longer passes through solid walls.
Added win detection so the game progresses to the next level when you
are on the green block. This is done by setting the direction to blank
so that the ball does not move in any direction when encountering an
object that is not passable. For example, I check whether the block to
the left of the current ball position is an x if the direction is west.
If the direction is to the east, then the block to the right will be
checked and the direction will be checked for stoppage.

Reset the position and paint of the ball when a new level is reached.

Painted the teleport block but not animated as that would require more
code.