. Changed movement detection to use try catch blocks instead of an if statement. It is more efficient and less confusing.
Added functionality to the teleport block.

This showed an issue where the ball does not move to the direction that
is intended but instead the block +1 in the direction the ball was
moving in. This is the reason that the ball skips the first block when
it loops around the array.