. Made sure that the windows cannot be resized.
Added a mouse listener to the two canvases instead of the JFrame. This
fixed the issue where the mousePressed method wouldn't be triggered.

Added an if statement where a brush variable would be changed based on
where on the 'pallet' the user clicked on.

The level would then be changed when the user clicks on a specific tile
(that tile would be changed based on which brush was selected).

Added a statement that draws the starting block if in the level editor
but does not when the user is playing the game (as that is where the
ball starts so showing the start position would only confuse the user).