. Made use of timers more efficient by only calling the run once instead of running it every time an arrow key is pressed.
Added a restart button but the ball won't move after the level is
restarted and a JOptionPane pops up when the restart method is called
because it utilizes the next level method.

Added more colision detection for the directional arrow keys but it
doesn't work.

Added some comments to clarify what the code does.

moved the drawing of the grid to after all the blocks so it overwrites
them and the weird glitch on the south and east sides doesn't occur.