. Changed the timer to trigger twice as fast for smoother movement.
Fixed the issue where restarting would cause the ball to not be able to
move. (.setFocusable()).

Changed the restart method to use it's own code instead of reusing
previous code that doesn't work as well.

Fixed collision detection on < and > blocks from the north and south.

Added a drop down for the user to select a level they would like to
edit. This replaced the system where the user typed in the name of the
file which was very error prone.

Changed the animation speed to keep is constant even though the timer is
moving twice as fast.