.
Renamed the level files so they would appear in the correct order should
one want to edit or play them in the custom level picker.

Changed the index of the starting element in the custom level picker so
that it would show the first one as the default option.

Centred the custom level picker that appears after the user beats a
custom level.

Made the restart method more efficient by using one for loop instead of
two.